resources/lambda/requirements.txt
boto3 
psycopg2-binary
aws_lambda_powertools